third_party/blink/web_tests/external/wpt/html/browsers/history/the-location-interface/location-prototype-setting-same-origin-domain.sub-expected.txt

This is a testharness.js-based test.
PASS Same-origin-domain prerequisite check: the original prototype is accessible
FAIL Same-origin-domain: setting the prototype to an empty object via Object.setPrototypeOf should throw a TypeError assert_throws: function "() => {
      Object.setPrototypeOf(target, newValue);
    }" did not throw
FAIL Same-origin-domain: setting the prototype to an empty object via __proto__ should throw a TypeError assert_throws: function "function() {
      target.__proto__ = newValue;
    }" did not throw
FAIL Same-origin-domain: setting the prototype to an empty object via Reflect.setPrototypeOf should return false assert_false: expected false got true
FAIL Same-origin-domain: the prototype must still be its original value assert_equals: expected object "[object Location]" but got object "[object Object]"
PASS Same-origin-domain: setting the prototype to its original value via Object.setPrototypeOf should not throw
PASS Same-origin-domain: setting the prototype to its original value via __proto__ should not throw
PASS Same-origin-domain: setting the prototype to its original value via Reflect.setPrototypeOf should return true
Harness: the test ran to completion.
